# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Production-Grade Web3 RFQ Market Maker API
  version: 0.1.0
  description: API draft for RFQ quote, submit, administrative quote control, health, and metrics flows. Every response includes an x-trace-id header for request correlation; structured error bodies also include the same traceId value. Clients may send a safe tr_-prefixed x-trace-id request header, which the gateway echoes after length and character validation; unsafe values fall back to a generated request id. Business and administrative routes require a scoped x-api-key in non-local deployments; only the key id and SHA-256 secret digest are stored server-side. Quotes belong to the authenticated stable principal, not an individual key id, so same-principal key rotation preserves access while cross-principal quote, settlement, and hedge lookups return the same 404 response as an absent resource; submit and PnL are scoped identically. The principal is off-chain authorization metadata and is never part of a public response or EIP-712 payload. Browser access is controlled by the RFQ_CORS_ALLOWED_ORIGINS allowlist, which only accepts normalized HTTP(S) URL origins without path, query, fragment, credentials, or wildcards. The gateway emits no-store cache control and baseline browser security headers on every response. Unknown routes and unsupported methods return the standard ErrorResponse shape.
servers:
  - url: http://localhost:3000
security:
  - ApiKeyAuth: []
paths:
...
  /admin/toxic-flow/scores/{chainId}/{user}:
    parameters:
      - name: chainId
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          maximum: 9007199254740991
      - name: user
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/Address"
    get:
      operationId: getToxicFlowScore
      summary: Read the latest dynamic toxic-flow score for one chain and user
      description: Returns the normalized key and latest versioned score. A null state means no dynamic score exists and base risk policy remains authoritative.
      x-required-scope: admin:read
      tags:
        - Administration
      responses:
        "200":
          description: Normalized score key and optional current state
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToxicFlowScoreLookup"
        "400":
          description: Invalid chain or user address
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing, malformed, invalid, or expired API key
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: API key lacks admin:read
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
            Retry-After:
              description: Seconds until the operator may retry.
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Toxic-flow score or rate-limit store unavailable
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateToxicFlowScore
      summary: Publish a versioned toxic-flow score for one chain and user
      description: Creates version 1 only with expectedVersion 0, then performs compare-and-swap updates and atomically appends immutable analyzer audit evidence.
      x-required-scope: admin:write
      tags:
        - Administration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToxicFlowScoreUpdate"
      responses:
        "200":
          description: Updated toxic-flow score state
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToxicFlowScoreState"
        "400":
          description: Invalid chain, user, or closed score update payload
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing, malformed, invalid, or expired API key
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: API key lacks admin:write
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: expectedVersion no longer matches the chain-user score state
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
            Retry-After:
              description: Seconds until the analyzer may retry.
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Toxic-flow score or rate-limit store unavailable
          headers:
            x-trace-id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Scoped key in keyId.secret form. Store only SHA-256(secret) in RFQ_API_KEY_CONFIG_JSON and rotate with overlapping key ids.
  headers:
    TraceId:
      description: Request correlation id attached to every response. ErrorResponse.traceId contains the same value for structured error bodies. Safe tr_-prefixed request header values may be echoed; unsafe values are replaced with a generated request id.
      schema:
        type: string
        pattern: "^tr_.+"
  schemas:
    Address:
      type: string
      pattern: "^0x[a-fA-F0-9]{40}$"
...
    SafeIdentifier:
      type: string
      minLength: 1
      maxLength: 128
      pattern: "^[A-Za-z0-9_:-]+$"
      description: Resource identifier containing only letters, numbers, underscore, colon, or hyphen.
...
    ToxicFlowScoreKey:
      type: object
      additionalProperties: false
      required:
        - chainId
        - user
      properties:
        chainId:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        user:
          $ref: "#/components/schemas/Address"
    ToxicFlowScoreState:
      type: object
      additionalProperties: false
      required:
        - chainId
        - user
        - scoreBps
        - postTradeDriftBps
        - sampleSize
        - windowSeconds
        - policyVersion
        - observedAt
        - version
        - updatedBy
        - updatedAt
      properties:
        chainId:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        user:
          $ref: "#/components/schemas/Address"
        scoreBps:
          type: integer
          minimum: 0
          maximum: 10000
        postTradeDriftBps:
          type: integer
          minimum: -10000
          maximum: 10000
        sampleSize:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        windowSeconds:
          type: integer
          minimum: 1
          maximum: 604800
        policyVersion:
          $ref: "#/components/schemas/SafeIdentifier"
        observedAt:
          type: string
          format: date-time
          pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
        version:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        updatedBy:
          type: string
          minLength: 1
          maxLength: 256
          pattern: "^[A-Za-z0-9_:-]+$"
        updatedAt:
          type: string
          format: date-time
          pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
    ToxicFlowScoreUpdate:
      type: object
      additionalProperties: false
      required:
        - scoreBps
        - postTradeDriftBps
        - sampleSize
        - windowSeconds
        - policyVersion
        - observedAt
        - expectedVersion
      properties:
        scoreBps:
          type: integer
          minimum: 0
          maximum: 10000
        postTradeDriftBps:
          type: integer
          minimum: -10000
          maximum: 10000
        sampleSize:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        windowSeconds:
          type: integer
          minimum: 1
          maximum: 604800
        policyVersion:
          $ref: "#/components/schemas/SafeIdentifier"
        observedAt:
          type: string
          format: date-time
          pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
        expectedVersion:
          type: integer
          minimum: 0
          maximum: 9007199254740991
    ToxicFlowScoreLookup:
      type: object
      additionalProperties: false
      required:
        - key
        - state
      properties:
        key:
          $ref: "#/components/schemas/ToxicFlowScoreKey"
        state:
          oneOf:
            - $ref: "#/components/schemas/ToxicFlowScoreState"
            - type: "null"
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - traceId
      properties:
        code:
          type: string
          enum:
            - INVALID_REQUEST
            - AUTHENTICATION_REQUIRED
            - AUTHORIZATION_DENIED
            - UNSUPPORTED_CHAIN
            - UNSUPPORTED_TOKEN
            - AMOUNT_TOO_SMALL
            - AMOUNT_TOO_LARGE
            - MARKET_DATA_UNAVAILABLE
            - ROUTING_UNAVAILABLE
            - PRICING_UNAVAILABLE
            - RISK_REJECTED
            - SIGNER_UNAVAILABLE
            - INVALID_SIGNATURE
            - QUOTE_STORE_UNAVAILABLE
            - QUOTE_NOT_FOUND
            - QUOTE_EXPIRED
            - QUOTE_ALREADY_USED
            - QUOTE_FAILED
            - QUOTE_PAUSED
            - QUOTE_CONTROL_CONFLICT
            - QUOTE_CONTROL_UNAVAILABLE
            - TOXIC_FLOW_SCORE_CONFLICT
            - TOXIC_FLOW_SCORE_UNAVAILABLE
            - HEDGE_NOT_FOUND
            - HEDGE_STORE_UNAVAILABLE
            - SETTLEMENT_EVENT_NOT_FOUND
            - SETTLEMENT_EVENT_STORE_UNAVAILABLE
            - PNL_STORE_UNAVAILABLE
            - SETTLEMENT_UNAVAILABLE
            - SETTLEMENT_REVERTED
            - SUBMIT_RESERVATION_UNAVAILABLE
            - RATE_LIMITED
            - RATE_LIMIT_UNAVAILABLE
            - INTERNAL_ERROR
        message:
          type: string
        traceId:
          type: string
          pattern: "^tr_.+"